requirements/local.txt
-r base.txt
flake8~=3.8.3
black~=19.10b0
isort[pyproject]~=4.3.21
mypy~=0.780
django-debug-toolbar~=2.2
coverage~=5.1
factory-boy~=2.12.0
django-silk~=4.0.1